+++ src/modules/auth/auth.controller.ts
@@ -1,5 +1,6 @@
 import { Body, Controller, Post, UsePipes, ValidationPipe } from '@nestjs/common';
 import { AuthService } from './auth.service';
+import { LoginDto } from './dto/login.dto';
 import { RegisterDto } from './dto/register.dto';
 
 @Controller('v1/auth')
@@ -15,5 +16,16 @@ export class AuthController {
   )
   register(@Body() dto: RegisterDto) {
     return this.authService.register(dto);
+  }
+
+  @Post('login')
+  @UsePipes(
+    new ValidationPipe({
+      whitelist: true,
+      forbidNonWhitelisted: true,
+    }),
+  )
+  login(@Body() dto: LoginDto) {
+    return this.authService.login(dto);
   }
 }